backend/requirements.txt
# FastAPI 和服务器
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.25

# 任务队列
celery==5.3.6
redis==5.0.1

# PDF 处理
pdf2docx==0.5.6
pdfplumber==0.10.3
PyPDF2==3.0.1
pikepdf==8.11.2
pdf2image==1.16.3
python-pptx==0.6.23

# 阿里云 OSS
oss2==2.18.4

# 支付
alipay-sdk-python==3.7.3

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 工具
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0
pillow==10.2.0

# 日志
structlog==24.1.0

# 开发工具
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
